laforge.lbr: repair PULSE_T1094NL footprint and symbol (solves OSM#3270)
... and OSM#3271. - The transformer symbol is made of two identical coils
and hence it is not possible to mark the "2:" side (pins 9-11) separately.
But the naming of this coil was changed in a way, that, when invoked, the
"2" of the subpart name "TR1-1:2" shows towards the pins 9-11.
Maybe we should evolve to a device symbol with two coils merged rather than
two identical but separated, then an individual naming could honour the
2:1 side and can not be disturbed by e.g. subsequent mirroring. -
Also repaired the name and value designators' layer association.